+++ deposit/src/main/java/com/dev_high/deposit/application/dto/DepositDto.java
@@ -3,6 +3,8 @@ package com.dev_high.deposit.application.dto;
 import com.dev_high.deposit.domain.Deposit;
 import com.dev_high.deposit.domain.DepositType;
 
+import java.math.BigDecimal;
+
 public class DepositDto {
     public record CreateCommand(
             String userId
@@ -16,16 +18,16 @@ public class DepositDto {
             String userId,
             String depositOrderId,
             DepositType type,
-            long amount
+            BigDecimal amount
     ) {
-        public static UsageCommand of(String userId, String depositOrderId, DepositType type, long amount) {
+        public static UsageCommand of(String userId, String depositOrderId, DepositType type, BigDecimal amount) {
             return new UsageCommand(userId, depositOrderId, type, amount);
         }
     }
     
     public record Info(
             String userId,
-            Long balance
+            BigDecimal balance
     ) {
         public static Info from(Deposit deposit) {
             return new Info(
