+++ deposit/src/main/java/com/dev_high/deposit/application/dto/DepositOrderDto.java
@@ -3,13 +3,14 @@ package com.dev_high.deposit.application.dto;
 import com.dev_high.deposit.domain.DepositOrder;
 import com.dev_high.deposit.domain.DepositOrderStatus;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public class DepositOrderDto {
     public record CreateCommand(
-            long amount
+            BigDecimal amount
     ) {
-        public static CreateCommand of(long amount) {
+        public static CreateCommand of(BigDecimal amount) {
             return new CreateCommand(amount);
         }
     }
@@ -26,7 +27,7 @@ public class DepositOrderDto {
     public record Info(
             String orderId,
             String userId,
-            long amount,
+            BigDecimal amount,
             DepositOrderStatus status,
             OffsetDateTime createdAt
     ) {
